Update deque value addition in InMemoryDB
Modified the value addition logic in the `InMemoryDB` class to extend the deque directly with original values instead of reversed values. This change alters the order of values stored in the deque.

Modify value addition in InMemoryDB deque

Changed the method of adding values to the deque in the
InMemoryDB class from extending with reversed values to
extending with values directly. This adjustment alters
the order of